Support endpoint detection in Python (#56)
* Fix Python API

+++ sherpa-ncnn/csrc/greedy-search-decoder.h
@@ -51,7 +51,7 @@ class GreedySearchDecoder : public Decoder {
     decoder_out_ = model_->RunDecoder(decoder_input_);
   }
 
-  void AcceptWaveform(int32_t sample_rate, const float *input_buffer,
+  void AcceptWaveform(float sample_rate, const float *input_buffer,
                       int32_t frames_per_buffer) override;
 
   void Decode() override;
